stress-test/requirements.txt
locust>=2.20.0
websocket-client>=1.7.0
